pyproject.toml
[project]
name = "iot-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.39.0",
    "influxdb-client>=1.45.0",
    "redis>=5.0.0",
]
